Initial Import and updates
