+++ P2P/plugin/communication/fermat-p2p-plugin-communication-network-node-bitdubai/src/main/java/com/bitdubai/fermat_p2p_plugin/layer/communications/network/node/developer/bitdubai/version_1/structure/data/node/request/ActorCatalogTransactionsMsjRequest.java
@@ -1,0 +1,18 @@
+/*
+ * @#ActorCatalogTransactionsMsjRequest.java - 2016
+ * Copyright bitDubai.com., All rights reserved.
+ * You may not modify, use, reproduce or distribute this software.
+ * BITDUBAI/CONFIDENTIAL
+ */
+package com.bitdubai.fermat_p2p_plugin.layer.communications.network.node.developer.bitdubai.version_1.structure.data.node.request;
+
+/**
+ * The Class <code>com.bitdubai.fermat_p2p_api.layer.all_definition.communication.commons.data.node.request.ActorCatalogTransactionsMsjRequest</code>
+ * <p/>
+ * Created by Roberto Requena - (rart3001@gmail.com) on 05/04/16.
+ *
+ * @version 1.0
+ * @since Java JDK 1.7
+ */
+public class ActorCatalogTransactionsMsjRequest {
+}
